Taxonomies are now ordered
In order to persist the terms ordering, the dictionary is annotated with
an ordering mapping.

The data structure is backwards compatible in such a way that when the
taxonomy is first edited (through-the-web or by importing a new VDEX
file), the ordering information is applied and will be used going
forward.